edp_mvp/requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.1.2
google-api-python-client==2.128.0
google-auth==2.29.0
pandas==2.2.2
